Simplify undo entries create/delete
